Disambiguation (scenario 2 and 3) implemented
There are some other things in this unrelated to disambiguation, but I'm
worried about some uncommitted depedencies and syncing up.

Issuing a learn command such as:

clotho.learn('test1 is fun', "clotho.say('it sure is');");

then putting in test1 and clicking submit does push the autocomplete,
but there appears to be some weirdness in how the client responds to it.
I think it probably doesn't display because of a loss of textfield
focus.