Added Trie to Mind
Try typing into the searchbar "mi...".  You'll get completions first
from the mind and then from the global.  Mindtest1 and mindtrap come
from the Mind.

The Mind's Trie is <String,String>.  The Global Trie is <String, Object>
where the object is some representation of a UUID and a list of
Sharables mapped to this synonym.

So, the mind's one is a 1:1 map, while the global one should be 1:n.

I also have changed the population of autocomplete results to grab the
key of the objects rather than the value since we are just returning the
autocompleted word here, not the uuid.