execution.subprocess: report API exceptions to subprocess
Before, if the subprocess made an API call that throws a Java exception,
then the entire run call unconditionally dies and the subprocess gets
cleaned up.

Now we send a special message to the subprocess when an API call
throws a Java exception. The user function then gets a special
exception thrown, which it can catch and recover.