Silencing of sloppiness
I've silenced out the sloppiness code in submit, as well as the methods
that are slaves of it.

I've tested:
clotho.get("org.clothocad.model.Trail");
In the search bar and it still returns data, which is right.

I've also debugged through:
pBca1020
which correctly errors out after being run in the scriptengine.

I also debugged through:
revcomp atcg
which similarly erred correctly.